Finish orderDb, some tests simple tests are missing...
