+++ src/test/java/secondWeek/Ordersystem/services/ProductServiceTest.java
@@ -1,0 +1,33 @@
+package secondWeek.Ordersystem.services;
+
+import org.junit.jupiter.api.Test;
+
+public class ProductServiceTest {
+
+ // Diese müssten noch geschrieben werden...
+
+ @Test
+ void addAndListProduct () {
+ }
+
+ @Test
+ void listProductsEmpty () {
+ }
+
+ @Test
+ void getProductById () {
+ }
+
+ @Test
+ void getProductByIdInvalid () {
+
+ }
+
+ @Test
+ void removeProductById () {
+ }
+
+ @Test
+ void removeProductByIdInvalid () {
+ }
+}
